% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareData.R
\name{compareData}
\alias{compareData}
\title{Comparing Two Prediction Models}
\usage{
compareData(df, model1, model2, x, metrics = c("rmse", "mae", "r2"))
}
\arguments{
\item{df}{A data frame to use for prediction and comparison.}

\item{model1}{First Model.}

\item{model2}{Second Model.}

\item{x}{Variable you want to compare the models predictions.}

\item{metrics}{Metrics to compare.}
}
\value{
A data frame comparing metrics
}
\description{
Compares Two fitted models with Root Mean Squared Error, Mean Absoulte Error, and R^2.
}
\examples{

# Example using iris dataset
flowers <- iris
split <- splitData(flowers, trainingRatio = 0.8, seed = NULL)
trainingOne <- split$training
testingOne <- split$testing

model1 <- lm(Sepal.Length ~ Petal.Length + Petal.Width, data = trainingOne)
model2 <- lm(Sepal.Length ~ Petal.Length + Petal.Width + Species, data = trainingOne)

results <- compareData(testingOne, model1, model2, x = "Sepal.Length")
results
}
