% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareGraph.R
\name{CompareGraph}
\alias{CompareGraph}
\alias{compareGraph}
\title{Comparing Models with Visuals}
\usage{
compareGraph(compareDataResults, plot = "metrics")
}
\arguments{
\item{compareDataResults}{A data frame that resulted from compareData}

\item{plot}{A string indicating the type of plot (metrics, residuals, actualvspredicted).}
}
\value{
Returns a ggplot2 comparison of two different models
}
\description{
Makes graphs to compare the compareData results and returns a ggplot2 graph
}
\examples{
flowers <- iris
split <- splitData(flowers, trainingRatio = 0.8, seed = 123)
trainingOne <- split$training
testingOne <- split$testing

model1 <- lm(Sepal.Length ~ Petal.Length + Petal.Width, data = trainingOne)
model2 <- lm(Sepal.Length ~ Petal.Length + Petal.Width + Species, data = trainingOne)

results <- compareData(testingOne, model1, model2, x = "Sepal.Length")

# Create and display a metrics comparison plot
comparison <- compareGraph(results, plot = "metrics")
comparison

}
