% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitData.R
\name{splitData}
\alias{splitData}
\title{Splits Data into Training and Testing Sets}
\usage{
splitData(df, trainingRatio = 0.8, seed = NULL)
}
\arguments{
\item{df}{A data frame}

\item{trainingRatio}{A ratio between 0 and 1 indicating the proportion of the dataset used for training (Default is 0.8).}

\item{seed}{Seed for reproducible results (null by default).}
}
\value{
A list with a training and testing dataframe:
\describe{
  \item{training}{The training subset of the dataframe.}
  \item{testing}{The testing subset of the dataframe.}
}
}
\description{
Takes in a dataframe, a training ratio, and a seed and returns a training and testing dataset in a list.
}
\examples{
# How to use splitData() function
flowers <- iris
head(flowers)

flowers <- splitData(flowers, trainingRatio = 0.8, seed = NULL)

# Use variables to extract the new dataframes
trainingOne <- flowers$training
testingOne <- flowers$testing

# Checking structure of the new data
str(trainingOne)
str(testingOne)

}
